+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## 1.1.0 (2023-10-07)
+* Support service account impersonation
+* Support the project and trace id variables
+* Update dependencies
 
 ## 1.0.0 (2023-06-16)